refactor(factoryinsight): merged kpis_methods and kpis_data into kpis
