+++ golang/cmd/factoryinsight/v2/models/dto-kpis.go
@@ -1,5 +1,7 @@
 package models
 
+import "time"
+
 type Kpi struct {
 	Id   int    `json:"id"`
 	Name string `json:"name"`
@@ -15,6 +17,20 @@ type GetKpisRequest struct {
 	AreaName           string `uri:"areaName" binding:"required"`
 	ProductionLineName string `uri:"productionLineName" binding:"required"`
 	WorkCellName       string `uri:"workCellName" binding:"required"`
+}
+
+type GetKpisDataRequest struct {
+	EnterpriseName     string `uri:"enterpriseName" binding:"required"`
+	SiteName           string `uri:"siteName" binding:"required"`
+	AreaName           string `uri:"areaName" binding:"required"`
+	ProductionLineName string `uri:"productionLineName" binding:"required"`
+	WorkCellName       string `uri:"workCellName" binding:"required"`
+	KpisMethod         string `uri:"kpisMethod" binding:"required"`
+}
+
+type KpiRequest struct {
+	From time.Time `form:"from" binding:"required"`
+	To   time.Time `form:"to" binding:"required"`
 }
 
 const (
